Auto detect slug, verbose name, fields, search and sorting
This commit allows you to write much less code when defining an AdvancedReport subclass.
It will automatically give itself a slug based on the xxxReport name (where xxx.lower() is applied).
It will also look into your model meta for verbose names.
Lastly, it checks if your model has been registered with the Django Admin. If this is the case, we
will steal the ``list_display``, ``sortable_fields`` and ``search_fields`` from it.