Add possibility to override the outputted HTML for report cells
Sometimes you want, for a specific report, transform a value of a certain type
(for example boolean) to something more friendly than the raw string represantation.
It was already possible on the field-level (with get_FOO_html) but this prevents
code duplication when you need it on more columns or reports.