Fix the ugly modal dialog scrollbar bug and reintroduce round corners
In order to enhance the visual experience, the nasty ugly scrollbar you see when
a dialog box pops up is removed.

Also, to go back to a more standard Bootstrap look and feel, the rounded corners are back
on panels. They looked a bit unnatural and unpolished, and it was clearly visible that
the previous change to rectangular corners was not done well (the background still had
rounded corners but the border not, which is just plain ugly...)