Add support for dialog modals as a response to _multiple actions.
See the NoModelReport for an example of this new feature in action.
You can go to the test website and click on "Examples".

It is now also possible to explicitly define a width from within the
action itself. Why is this useful? Sometimes you just don't want a dialog
to take up the whole width of the page (which is the default), because
you maybe have some very small content. Also for this, check the
NoModelReport example on the Examples page :-)