Bootstrap templates for the non-AngularJS version
Added some templates based on bootstrap. The idea here is to replace the current defaults,
which are honestly speaking as bad as can be. They even depend on "helpdesk/base.html", which
is not included. It is still an artifact of the past. The new templates will used as the
new default when using new-style reports.