Do not show the row limit selector by default
A long time ago, someone added this feature for the helpdesk of Poland.
It is not fully compatible with other feautures of Advanced Reports,
therefore I made this an optional setting which is by default off.

The Polish people are not impacted, as they have their own fork of
django-advanced-reports.